+++ packages/boilerplate-generator/package.js
@@ -12,7 +12,8 @@ Package.onUse(function (api) {
   // spacebars compiler rather than letting the 'templating' package (which
   // isn't fully supported on the server yet) handle it. That also means that
   // they don't contain the outer "<template>" tag.
-  api.addFiles(['boilerplate_web.browser.html',
-                 'boilerplate_web.cordova.html'],
-                 'server', {isAsset: true});
+  api.addAssets([
+    'boilerplate_web.browser.html',
+    'boilerplate_web.cordova.html'
+  ], 'server');
 });