+++ packages/reload-safetybelt/package.js
@@ -6,11 +6,11 @@ Package.describe({
 Package.onUse(function (api) {
   api.use("webapp", "server");
   api.addFiles("reload-safety-belt.js", "server");
-  api.addFiles("safetybelt.js", "server", { isAsset: true });
+  api.addAssets("safetybelt.js", "server");
 });
 
 Package.onTest(function (api) {
-  api.addFiles("safetybelt.js", "server", { isAsset: true });
+  api.addAssets("safetybelt.js", "server");
   api.use(["reload-safetybelt", "tinytest", "http", "webapp", "underscore"]);
   api.addFiles("reload-safety-belt-tests.js", "server");
 });